backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "LibreM FastAPI backend service"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.109.0",
    "sqlalchemy>=2.0.25",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.5.3",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
]
